volleyball-matches-scraper/requirements.txt
ipykernel
nest_asyncio
fake-useragent
stem
pytest-playwright
beautifulsoup4
tqdm